+++ lib/dynamic_cache.dart
@@ -3,3 +3,4 @@ library dynamic_cache;
 
 export 'package:dynamic_cache/sources/dynamic_cache.dart';
 export 'package:dynamic_cache/sources/dynamic_cache_provider.dart';
+export 'package:flutter/material.dart';