+++ lib/sources/dynamic_cache_provider.dart
@@ -2,25 +2,30 @@ import 'package:dynamic_cache/sources/dynamic_cache.dart';
 import 'package:flutter/widgets.dart';
 
 /// Extension for ease access to the cache.
-extension DynamicCacheExtension on BuildContext {
+extension DynamicCacheCacheExtension on BuildContext {
   DynamicCache get cache => DynamicCacheProvider.of(this);
 }
 
-/// Wrap your app with [DynamicCacheProvider] for
-/// access to the cache instance on context.cache .
+extension AsyncSnapshotCacheExtension on AsyncSnapshot<Map<String, dynamic>> {
+  T fromCache<T>(String key) {
+    return requireData[key] as T;
+  }
+}
+
+/// Wrap your app with DynamicCacheProvider
+/// for access to the cache instance by context.cache...
 class DynamicCacheProvider extends InheritedNotifier<DynamicCache> {
   const DynamicCacheProvider({
+    required this.cache,
     required super.child,
     super.key,
-
-    /// Note: specifying the cache is optional, as it will
-    /// create one if it does not receive the parameter.
-    DynamicCache cache = DynamicCache.instance,
   }) : super(notifier: cache);
 
-  static DynamicCache of(BuildContext context, {bool listen = true}) {
-    final inherited =
-        context.dependOnInheritedWidgetOfExactType<DynamicCacheProvider>()!;
+  final DynamicCache cache;
+
+  static DynamicCache of(BuildContext context) {
+    final inherited = context
+        .dependOnInheritedWidgetOfExactType<DynamicCacheProvider>()!;
     return inherited.notifier!;
   }
 }