+++ test/dynamic_cache_test.dart
@@ -5,22 +5,6 @@ void main() {
   late List<String> names;
   const instance = DynamicCache.instance;
 
-  test('Test instances of DynamicCache', () {
-    final instanceA = DynamicCache();
-    final instanceB = DynamicCache();
-    expect(
-      instanceA == instanceB,
-      isA<bool>().having((value) => value, 'Expect false', false),
-    );
-
-    const instanceC = DynamicCache.instance;
-    const instanceD = DynamicCache.instance;
-    expect(
-      instanceC == instanceD,
-      isA<bool>().having((value) => value, 'Expect true', true),
-    );
-  });
-
   setUp(() {
     names = const [
       'John',
@@ -34,7 +18,7 @@ void main() {
   });
 
   test('Add and remove items from memory cache', () async {
-    instance.create(
+    instance.add(
       'names',
       names,
       expiration: const Duration(seconds: 5),
@@ -54,4 +38,5 @@ void main() {
       isA<bool>().having((value) => value, 'Expect empty list', false),
     );
   });
+
 }